moved common/models/element to types/element
