+++ apps/studio/electron/main/code/diff/index.ts
@@ -4,7 +4,7 @@ import { readFile } from '../files';
 import { parseJsxFile, removeSemiColonIfApplicable } from '../helpers';
 import { transformAst } from './transform';
 import type { CodeDiff, CodeDiffRequest } from '@onlook/types/code';
-import type { TemplateNode } from '/common/models/element/templateNode';
+import type { TemplateNode } from '@onlook/types/element/templateNode';
 
 interface RequestsByPath {
     templateToCodeDiff: Map<TemplateNode, CodeDiffRequest>;
