+++ apps/studio/electron/main/code/templateNode.ts
@@ -2,7 +2,7 @@ import traverse from '@babel/traverse';
 import type t from '@babel/types';
 import { readCodeBlock } from '.';
 import { parseJsxFile } from './helpers';
-import type { TemplateNode, TemplateTag } from '/common/models/element/templateNode';
+import type { TemplateNode, TemplateTag } from '@onlook/types/element/templateNode';
 
 export async function getTemplateNodeChild(
     parent: TemplateNode,
