+++ apps/studio/electron/preload/webview/elements/index.ts
@@ -1,7 +1,7 @@
 import { getDomElement } from './helpers';
 import { EditorAttributes } from '@onlook/types/constants';
 import { getUniqueSelector } from '/common/helpers';
-import type { DomElement } from '/common/models/element';
+import type { DomElement } from '@onlook/types/element';
 
 export const getSelectorAtLoc = (x: number, y: number): string => {
     const el = getDeepElement(x, y) || document.body;
