+++ apps/studio/electron/preload/webview/elements/text.ts
@@ -1,7 +1,7 @@
 import { publishEditText } from '../events/publish';
 import { getDomElement, getImmediateTextContent, restoreElementStyle } from './helpers';
 import { EditorAttributes } from '@onlook/types/constants';
-import type { TextDomElement } from '/common/models/element';
+import type { TextDomElement } from '@onlook/types/element';
 
 export function editTextBySelector(selector: string, content: string): TextDomElement | null {
     const el: HTMLElement | null = document.querySelector(selector);
