+++ apps/studio/src/lib/editor/engine/code/helpers.ts
@@ -1,7 +1,7 @@
 import { twMerge } from 'tailwind-merge';
 import { CssToTailwindTranslator } from '/common/helpers/twTranslator';
 import type { CodeDiffRequest } from '@onlook/types/code';
-import type { TemplateNode } from '/common/models/element/templateNode';
+import type { TemplateNode } from '@onlook/types/element/templateNode';
 
 export async function getOrCreateCodeDiffRequest(
     templateNode: TemplateNode,
