+++ apps/studio/src/lib/editor/engine/copy/index.ts
@@ -10,7 +10,7 @@ import type {
     ActionTarget,
     InsertElementAction,
 } from '@onlook/types/actions';
-import type { WebViewElement } from '/common/models/element';
+import type { WebViewElement } from '@onlook/types/element';
 
 export class CopyManager {
     copied: {
