+++ apps/studio/src/lib/editor/engine/element/index.ts
@@ -3,7 +3,7 @@ import { makeAutoObservable } from 'mobx';
 import type { EditorEngine } from '..';
 import { escapeSelector } from '/common/helpers';
 import type { RemoveElementAction } from '@onlook/types/actions';
-import type { DomElement, WebViewElement } from '/common/models/element';
+import type { DomElement, WebViewElement } from '@onlook/types/element';
 
 export class ElementManager {
     private hoveredElement: WebViewElement | undefined;
