+++ apps/studio/src/lib/editor/engine/text/index.ts
@@ -2,7 +2,7 @@ import type { WebviewTag } from 'electron';
 import jsStringEscape from 'js-string-escape';
 import type { EditorEngine } from '..';
 import { escapeSelector } from '/common/helpers';
-import type { DomElement, TextDomElement } from '/common/models/element';
+import type { DomElement, TextDomElement } from '@onlook/types/element';
 
 export class TextEditingManager {
     isEditing = false;
