+++ apps/studio/src/routes/editor/EditPanel/StylesTab/single/TailwindInput.tsx
@@ -5,7 +5,7 @@ import { observer } from 'mobx-react-lite';
 import { useEffect, useRef, useState } from 'react';
 import { MainChannels } from '@onlook/types/constants';
 import type { CodeDiffRequest } from '@onlook/types/code';
-import type { TemplateNode } from '/common/models/element/templateNode';
+import type { TemplateNode } from '@onlook/types/element/templateNode';
 import { Icons } from '@onlook/ui/icons';
 
 const TailwindInput = observer(() => {
