+++ apps/studio/src/routes/editor/LayersPanel/Tree/TreeRow.tsx
@@ -1,6 +1,6 @@
 import type { RefObject } from 'react';
 import type { NodeApi } from 'react-arborist';
-import type { LayerNode } from '/common/models/element/layers';
+import type { LayerNode } from '@onlook/types/element/layers';
 
 interface TreeRowProps {
     node: NodeApi<LayerNode>;
