+++ apps/studio/src/routes/editor/RightClickMenu/index.tsx
@@ -11,8 +11,8 @@ import { cn } from '@onlook/ui/utils';
 import { observer } from 'mobx-react-lite';
 import { useEffect, useState } from 'react';
 import { Hotkey } from '/common/hotkeys';
-import type { WebViewElement } from '/common/models/element';
-import type { TemplateNode } from '/common/models/element/templateNode';
+import type { WebViewElement } from '@onlook/types/element';
+import type { TemplateNode } from '@onlook/types/element/templateNode';
 import { Icons } from '@onlook/ui/icons';
 
 interface RightClickMenuProps {
