+++ packages/types/src/element/templateNode.ts
